chore: move SchemaWithDomain to cells crate and make pub
