+++ tiledb/api/src/query/strategy/tests.rs
@@ -1,25 +1,17 @@
+use cells::strategy::SchemaWithDomain;
 use cells::write::strategy::{WriteParameters, WriteSequenceParameters};
 use cells::write::{DenseWriteInput, SparseWriteInput, WriteSequence};
 use proptest::prelude::*;
+use tiledb_common::Datatype;
 use tiledb_common::array::schema::EnumerationKey;
 use tiledb_common::array::{
     CellOrder, TileOrder, dimension::DimensionConstraints,
 };
-use tiledb_common::datatype::physical::BitsKeyAdapter;
 use tiledb_common::metadata::Value;
-use tiledb_common::query::condition::strategy::{
-    Parameters as QueryConditionParameters, QueryConditionField,
-    QueryConditionSchema,
-};
-use tiledb_common::query::condition::{
-    EqualityOp, Field as ASTField, SetMembers,
-};
-use tiledb_common::range::{
-    NonEmptyDomain, Range, SingleValueRange, VarValueRange,
-};
-use tiledb_common::{Datatype, set_members_go};
+use tiledb_common::query::condition::strategy::Parameters as QueryConditionParameters;
+use tiledb_common::range::{NonEmptyDomain, Range};
 use tiledb_pod::array::attribute::{AttributeData, FillData};
-use tiledb_pod::array::schema::{FieldData as SchemaField, SchemaData};
+use tiledb_pod::array::schema::SchemaData;
 use tiledb_pod::array::{DimensionData, DomainData, EnumerationData};
 use uri::TestArrayUri;
 
@@ -521,156 +513,6 @@ fn instance_query_condition(
     }
 
     Ok(())
-}
-
-struct SchemaWithDomain {
-    fields: Vec<FieldWithDomain>,
-}
-
-impl SchemaWithDomain {
-    pub fn new(schema: Rc<SchemaData>, cells: &Cells) -> Self {
-        Self {
-            fields: cells
-                .domain()
-                .into_iter()
-                .map(|(f, domain)| FieldWithDomain {
-                    schema: Rc::clone(&schema),
-                    field: schema.field(f).unwrap(),
-                    domain,
-                })
-                .collect::<Vec<_>>(),
-        }
-    }
-}
-
-struct FieldWithDomain {
-    schema: Rc<SchemaData>,
-    field: SchemaField,
-    domain: Option<Range>,
-}
-
-impl QueryConditionSchema for SchemaWithDomain {
-    fn fields(&self) -> Vec<&dyn QueryConditionField> {
-        self.fields
-            .iter()
-            .map(|f| f as &dyn QueryConditionField)
-            .collect::<Vec<_>>()
-    }
-}
-
-impl QueryConditionField for FieldWithDomain {
-    fn name(&self) -> &str {
-        self.field.name()
-    }
-
-    fn equality_ops(&self) -> Option<Vec<EqualityOp>> {
-        match self.field {
-            SchemaField::Dimension(_) => None,
-            SchemaField::Attribute(ref a) => {
-                if let Some(edata) = self
-                    .schema
-                    .enumeration(EnumerationKey::AttributeName(&a.name))
-                {
-                    if !ASTField::is_allowed_type(
-                        edata.datatype,
-                        edata.cell_val_num.unwrap_or(CellValNum::single()),
-                    ) {
-                        // only null test allowed for these
-                        Some(vec![])
-                    } else if matches!(edata.ordered, Some(true)) {
-                        // anything goes
-                        None
-                    } else {
-                        Some(vec![EqualityOp::Equal, EqualityOp::NotEqual])
-                    }
-                } else if !ASTField::is_allowed_type(
-                    a.datatype,
-                    a.cell_val_num.unwrap_or(CellValNum::single()),
-                ) {
-                    // only null test allowed for these
-                    Some(vec![])
-                } else {
-                    None
-                }
-            }
-        }
-    }
-
-    fn domain(&self) -> Option<Range> {
-        #[allow(clippy::collapsible_if)]
-        if let SchemaField::Attribute(ref a) = self.field {
-            if let Some(edata) = self
-                .schema
-                .enumeration(EnumerationKey::AttributeName(&a.name))
-            {
-                // query condition domain is in terms of the enumerated values,
-                // not the attribute values domaion (which are indexes into the enumerated values)
-                let members = edata.query_condition_set_members()?;
-                return Some(set_members_go!(
-                    members,
-                    _DT,
-                    ref members,
-                    {
-                        let min = *members.iter().min()?;
-                        let max = *members.iter().max()?;
-                        Range::Single(SingleValueRange::from(min..=max))
-                    },
-                    {
-                        let min = *members.iter().map(BitsKeyAdapter).min()?.0;
-                        let max = *members.iter().map(BitsKeyAdapter).max()?.0;
-                        Range::Single(SingleValueRange::from(min..=max))
-                    },
-                    {
-                        let min = members.iter().min()?.clone();
-                        let max = members.iter().max()?.clone();
-                        Range::Var(VarValueRange::from((
-                            min.into_bytes().into_boxed_slice(),
-                            max.into_bytes().into_boxed_slice(),
-                        )))
-                    }
-                ));
-            }
-        }
-
-        // see query_ast.cc
-        if matches!(
-            self.field.datatype(),
-            Datatype::Any
-                | Datatype::StringUtf16
-                | Datatype::StringUtf32
-                | Datatype::StringUcs2
-                | Datatype::StringUcs4
-                | Datatype::Blob
-                | Datatype::GeometryWkb
-                | Datatype::GeometryWkt
-        ) {
-            None
-        } else if matches!(self.domain, Some(Range::Single(_)))
-            || (matches!(
-                self.field.datatype(),
-                Datatype::StringAscii | Datatype::StringUtf8
-            ) && matches!(
-                self.field.cell_val_num(),
-                None | Some(CellValNum::Var)
-            ))
-        {
-            self.domain.clone()
-        } else {
-            None
-        }
-    }
-
-    fn set_members(&self) -> Option<SetMembers> {
-        match self.field {
-            SchemaField::Dimension(_) => None,
-            SchemaField::Attribute(ref a) => {
-                let edata = self
-                    .schema
-                    .enumeration(EnumerationKey::AttributeName(&a.name))?;
-                edata.query_condition_set_members()
-            }
-        }
-    }
 }
 
 fn strat_query_condition() -> impl Strategy<
